% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted rand index}
\usage{
adjusted_rand_index(pred_partition, true_partition)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{true_partition}{"true" partition, aka ground truth}
}
\value{
the adjusted rand index of the partitions
}
\description{
External metric.
}
