% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{analyze_survival}
\alias{analyze_survival}
\title{Plots the survival curves and compute the Cox p-value}
\usage{
analyze_survival(survival_time, death_status, patient_partition,
  return = FALSE, no_plot = FALSE)
}
\arguments{
\item{survival_time}{the follow-up time of the patients.}

\item{death_status}{integer, 0 = alive, 1 = dead}

\item{return}{logical, TRUE to return the cox p-value}

\item{no_plot}{logical, set to TRUE to avoid plotting and return
the cox p-value.}

\item{patients_partition}{a partition of the patients}
}
\value{
the cox p-value (if return or no_plot is TRUE).
}
\description{
Plots the survival curves and compute the Cox p-value using the package
\href{https://cran.r-project.org/web/packages/survival/index.html}
{survival}.
}
\seealso{
\code{\link[survival]{Surv}}, \code{\link[survival]{coxph}}
}
