% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ClinicalBoxplot}
\alias{ClinicalBoxplot}
\title{Box-plot of clinical features for a given partition of the samples.}
\usage{
ClinicalBoxplot(clinical.feature, partition, dot = FALSE,
  remove.outliers = FALSE, feature.name = "clinical.feature",
  group.name = "group")
}
\arguments{
\item{clinical.feature}{a numeric vector with a value for each patients.}

\item{partition}{an integer (or factor) vector indicating the group of each patient.}

\item{dot}{logical, to plot the points on the box-plot}

\item{remove.outliers}{logical, to only plot data between quantile 0.1 and 0.9 (FALSE by default)}

\item{feature.name}{the name of the clinical feature plotted.}

\item{group.name}{the name of the group in the partition, for the plot.}
}
\description{
Plot the box-plot of the given clinical features, splitting data according to the given partition.
}
