% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{consensus_spectral_clustering}
\alias{consensus_spectral_clustering}
\title{Apply consensus clustering on affinity matrix.}
\usage{
consensus_spectral_clustering(affinity_matrix, cluster_number_max = 5,
  reps = 10, pItem = 0.8, pFeature = 1,
  title = "Consensus spectral clustering", innerLinkage = "average",
  finalLinkage = "average", distance = "pearson", ml = NULL,
  tmyPal = NULL, seed = NULL, plot = NULL, writeTable = FALSE,
  weightsItem = NULL, weightsFeature = NULL, verbose = F,
  corUse = "everything")
}
\arguments{
\item{affinity_matrix}{affinity matrix (e.g. produced by snf or anf)}

\item{cluster_number_max}{integer. The maximum cluster number to evaluate.}

\item{reps}{Number of subsamples evaluated.}

\item{pItem}{numeric, proportion of items to sample.}

\item{pFeature}{numeric, proportion of features to sample.}

\item{title}{character for output directory. Directory is created only
if plot is not NULL or writeTable is TRUE. This title can be an abosulte
or relative path.}

\item{innerLinkage}{heirarchical linkage method for subsampling.}

\item{finalLinkage}{heirarchical linkage method for consensus matrix.}

\item{ml}{optional. prior result, if supplied then only do graphics and
tables.}

\item{tmyPal}{optional character vector of colors for consensus matrix}

\item{seed}{optional numerical value. sets random seed for reproducible
results.}

\item{plot}{character value. NULL - print to screen, 'pdf', 'png', 'pngBMP'
for bitmap png, helpful for large datasets.}

\item{writeTable}{logical value. TRUE - write ouput and log to csv.}

\item{weightsItem}{optional numerical vector. weights to be used for sampling
items.}

\item{weightsFeature}{optional numerical vector. weights to be used for
sampling features.}

\item{verbose}{boolean. If TRUE, print messages to the screen to indicate
progress. This is useful for large datasets.}

\item{corUse}{optional character value. specifies how to handle missing data
in correlation distances 'everything','pairwise.complete.obs',
'complete.obs' see cor() for description.}
}
\value{
results of ConsensusClusterPlus & calcICL.
}
\description{
Use spectral clustering as the clustering function of consensus clustering on
the provided affinity matrix.
}
\details{
It uses \code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}.
ConsensusClusterPlus implements the Consensus Clustering algorithm of Monti,
et al (2003). The function will subsamples the affinity matrix according
to pItem, pFeature, weightsItem, and weightsFeature, and clusters the data
into 2 to maxK clusters using spectral clustering.

It will also compute the item consensus results using
\code{\link[ConsensusClusterPlus]{calcICL}}. For more informations, see the
documentation of the original package!
}
