% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{DataDistribution}
\alias{DataDistribution}
\title{Plot the distribution of the features}
\usage{
DataDistribution(data)
}
\arguments{
\item{data}{a feature matrix with samples in rows and features in columns}
}
\description{
The function plots the distribution of the mean, the variance and the median absolute deviation
of the features.
}
