% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{dunn_index}
\alias{dunn_index}
\title{Dunn index}
\usage{
dunn_index(pred_partition, distance_matrix)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{distance_matrix}{a distance matrix}
}
\value{
the Dunn index of the partitions
}
\description{
Dunn index
}
