% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entropy}
\alias{entropy}
\title{Return entropy of vector}
\usage{
entropy(x)
}
\arguments{
\item{x}{an integer vector}
}
\value{
The entropy of x
}
\description{
Essentially number of bits required to specify
Supporting function for metrics below
}
