% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{evaluation}
\alias{evaluation}
\title{Evaluate parameters}
\usage{
evaluation(data_list, method, grid, grid_row, metric = NULL,
  true_partition = NULL, return_res = FALSE)
}
\arguments{
\item{data_list}{a list of data matrices with continuous data of format
samples x features (with the same number of samples).}

\item{method}{a string being the name of the built-in method to be used or a
'method list' with the same format as the built-in ones.}

\item{grid}{the grid generated in \code{tuning()}.}

\item{grid_row}{the row indicating the set of parameters to be tested in the grid}

\item{metric}{a string being the name of the built-in metric to be used or a
'metric list' with the same format as the built-in ones. If the chosen
metric is internal, be sure to choose a metric corresponding to the kind of
data returned by the chosen method.}

\item{true_partition}{a factor or integer vector being a classification of the
patients. It will be used with external metrics.}

\item{return_res}{logical. To return the result of the method instead of the value of the metric.}
}
\value{
the value of the metric or the result of the method if \code{return_res} is TRUE
}
\description{
Evaluate a set of parameters for a given method with a given metric.
}
