% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{filter_flagged_data}
\alias{filter_flagged_data}
\title{Reduce a sample-featyure matrix to suitable data}
\usage{
filter_flagged_data(m, flag_fxn = is.na)
}
\arguments{
\item{m}{a sample-feature matrix}

\item{flag_fxn}{a boolean function that detects problematic data}
}
\value{
the matrix with all rows and columns that contain problematic data removed
}
\description{
Reduce a sample-featyure matrix to suitable data
}
\examples{

}
