% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{flag_data}
\alias{flag_data}
\title{Simple quality control on sample-feature matrices}
\usage{
flag_data(m, flag_fxn = is.na, filter = FALSE, thresh = c(0.2, 0.2),
  plot = FALSE, print = TRUE)
}
\arguments{
\item{m}{A sample-feature matrix.}

\item{flag_fxn}{A boolean function that detects problematic data.}

\item{filter}{logical wether to filter the data. The rows (resp. columns)
with a percentage of flagged data greater than \code{thresh[1]}
(resp. \code{thresh[2]}) are removed.}

\item{thresh}{threshold vector: first for rows, second for columns.}

\item{plot}{logical, to plot the histogram of the percentage of flagged
values per rows or columns.}

\item{print}{logical, to print informations on the number of flagged elements
in the data.}
}
\value{
The filtered data if filter is TRUE. Nothing otherwise.
}
\description{
Many methods have restrictions on the data they can use (most notably
missing data). This is a simple function for assessing the quality of
input data. By default it looks for missing (NA) data although the
the trigger condition can be modified.
}
