% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_data.R
\docType{data}
\name{gbm_data}
\alias{gbm_data}
\title{Glioblastoma multiforme (GBM) multi-omics data}
\format{The dataset contains two lists:
\itemize{
\item $data_list: a list of 3 features matrix with patients in rows
and features in columns:
\itemize{
\item $mRNA: mRNA expression
\item $methylation: methylation profile
\item $miRNA: micro RNA expression
}
\item $survival: the survival data of the patients that contains:
\itemize{
\item $PatientID: ID of TCGA patients
\item $Survival: survival time of the patients
\item $Death: death status (0=alive, 1=dead)
}
}}
\usage{
gbm_data
}
\description{
The dataset comes from The Cancer Genome Atlas:
\url{https://cancergenome.nih.gov/}.
}
\details{
The dataset is curated level-three, see:
\url{https://cancergenome.nih.gov/abouttcga/aboutdata/datalevelstypes}.
}
\keyword{datasets}
