% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_multi_moclust}
\alias{generate_multi_moclust}
\title{Generate simulated data}
\usage{
generate_multi_moclust(data_support, sd_signal = 0.5, sparse = FALSE,
  percent_sparsity)
}
\arguments{
\item{data_support}{a list of features matrices}

\item{sd_signal}{the standard deviation of the signal in each cluster,
see 10.1021/acs.jproteome.5b00824, try 1, 0.5 or 0.2 for good to bad signal
to noise ratio}

\item{sparse}{logical, to create sparsity in singular vector of Xsim,
see 10.1021/acs.jproteome.5b00824}

\item{percent_sparsity}{a vector (value for each data type) indicating the
percentage of sparsity for the singular vector V of Xsim. See
10.1021/acs.jproteome.5b00824}
}
\value{
a list of simulated structured data, the corresponding partition
(partition_tot) and the partition per layer (partitions).
}
\description{
Generate simulated data as in 10.1021/acs.jproteome.5b00824
}
