% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_multi_structured}
\alias{generate_multi_structured}
\title{Generate synthetic data based on a list of features matrics.}
\usage{
generate_multi_structured(data_support, structure_type = c("moclus"),
  ...)
}
\arguments{
\item{data_support}{A list of features matrices used to generate the
data}

\item{structure_type}{The type of structure: "moclus" only.}

\item{...}{Parameters for the method of generation. See Details.}
}
\value{
A list of synthetic features matrix with the same dimension than the
\code{data_support}
}
\description{
Generate a structured list of data with 4 clusters.
}
\details{
For the moment, only one method of generation is provided:
\itemize{
\item "moclus" is inspired from
moCluster: Identifying Joint Patterns Across Multiple Omics Data Sets, DOI:
10.1021/acs.jproteome.5b00824. The single parameter is: sd_signal which is
the standard deviation of the signal used for the creation of the
structure. Default value is 1. The higher sd_signal is, the higher the
signal to noise ratio is. See the article for more information.
}
}
