% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_multi_unstructured}
\alias{generate_multi_unstructured}
\title{Generate a list of synthetic data matrices for validation}
\usage{
generate_multi_unstructured(type = c("gaussian", "uniform"), n_samples,
  n_features, n_layers)
}
\arguments{
\item{type}{The type of synthetic data generated, "gaussian" or "uniform".}

\item{n_samples}{The number of samples.}

\item{n_features}{integer vector. The number of features per layer.}

\item{n_layers}{The number of layers (features matrix) to
be generated.}
}
\value{
a list of unstructured matrix of the type selected in \code{type}
(gaussian by default).
}
\description{
Generate a list of unstructured data matrices of the desired dimensions.
}
\details{
Two types of distribution are available:
\itemize{
\item "gaussian": generates a matrix with each patient having features with
the same gaussian distribution.
\item "uniform" generates a matrix with each patient having features with
the same uniform distribution.
}
}
