% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_partitions}
\alias{generate_partitions}
\title{Generate a partitions matrix}
\usage{
generate_partitions(n_samples, n_layers, n_cluster = 4)
}
\arguments{
\item{n_samples}{The number of samples}

\item{n_layers}{The number of layers}

\item{n_cluster}{The number of cluster.}
}
\value{
A list:
\itemize{
\item $partition_tot: the intersection of the partitions
\item $partitions: a matrix with rows representing the partition for each
layer.
}
}
\description{
Used in generate_multi_structured methods to create partitions for each
layer. The intersection of these partitions are a global partition.
}
