% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_single_moclust}
\alias{generate_single_moclust}
\title{Generate simulated data}
\usage{
generate_single_moclust(X, partition, sd_signal = 0.5, sparse = FALSE,
  percent_sparsity = 0.3)
}
\arguments{
\item{X}{A feature matrix, rows are patients, columns are features.}

\item{partition}{A partition of the samples.}

\item{sd_signal}{The standard deviation of the signal in each cluster.}

\item{sparse}{logical, to create sparsity in singular vector of Xsim.}

\item{percent_sparsity}{Percentage of sparsity in singular vector of Xsim.}
}
\value{
A simulated structured matrix
}
\description{
Generate simulated data as in 10.1021/acs.jproteome.5b00824 (DOI).
}
\details{
Considering a partition of the samples, the data will be modified to have a
structure corresponding to this partition.
}
\seealso{
DOI: 10.1021/acs.jproteome.5b00824
}
