% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{GenerateSynthData}
\alias{GenerateSynthData}
\title{Generate 'multi-omic' synthetic data matrices  for validation}
\usage{
GenerateSynthData(type = c("gaussian", "uniform", "structured"),
  n_samples = NULL, n_layers = NULL, n_features = NULL,
  support.data.list = NULL, separation = 2)
}
\arguments{
\item{type}{type of data, "gaussian" is a homogeneous group of patients, made with the gaussian distribution. "structured" is made as the
simulated data is created in 10.1021/acs.jproteome.5b00824}

\item{n_samples}{the number of samples (rows)}

\item{n_layers}{integer. The number of layers (features matrix) to be generated}

\item{n_features}{a list of the numbers of features for each matrix}

\item{support.data.list}{a list of features matrix that are used to generate the data}

\item{separation}{double. The value of separation between clusters: the space between
the means of the normal distributions used to generate the structure of the synthetic data.}
}
\value{
'multi-omic' synthetic data matrices for validation
}
\description{
For validation and testing multi-omic / integrative methods, we need
}
