% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{Generator_Support_based}
\alias{Generator_Support_based}
\title{Generate synthetic data}
\usage{
Generator_Support_based(data.support, structure.type = c("basic", "moClus"),
  separation = 2)
}
\arguments{
\item{data.support}{a list of features matrix that are used to generate the data}

\item{structure.type}{the type of structure, "moClus" is inspired from
moCluster: Identifying Joint Patterns Across Multiple Omics Data Sets,
DOI: 10.1021/acs.jproteome.5b00824}

\item{separation}{double. The value of separation between clusters: the space between
the means of the normal distributions used to generate the structure of the synthetic data.}
}
\value{
a list of synthetic features matrix with the same dimension than the \code{data.support}
}
\description{
Generate synthetic data with a known structure based on a real dataset provided by the user
}
