% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{Generator_unstructured}
\alias{Generator_unstructured}
\title{Generate a list of synthetic data matrix for validation}
\usage{
Generator_unstructured(type = c("gaussian", "uniform"), n_samples, n_features,
  n_layers)
}
\arguments{
\item{type}{type of synthetic data generated : "gaussian" generates a matrix
with each patients having features with the same gaussian distribution
and "uniform" generates a matrix with each patients having features with
the same uniform distribution.}

\item{n_layers}{integer. The number of layers (features matrix) to be generated.}

\item{dims}{a vector of 2 integers, which are the dimensions of the matrix to
be generated, being the number of samples and features respectively}
}
\value{
a list of \code{n_layers} synthetic matrix of the type selected in \code{type}
(gaussian by default) with dimensions \code{dims}.
}
\description{
Generate a list of synthetic data matrix for validation
}
