% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetMethodInfo}
\alias{GetMethodInfo}
\title{GetMethodInfo}
\usage{
GetMethodInfo(method = NULL)
}
\arguments{
\item{method}{the name of the built-in method to be used}
}
\value{
a method list
}
\description{
GetMethodInfo
}
