% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetMetric}
\alias{GetMetric}
\title{Load built-in metrics}
\usage{
GetMetric(metric = NULL, extract = TRUE)
}
\arguments{
\item{metric}{a character. The metric to be used.}
}
\value{
a metric list. $name is the name used to access the metric with GetMetric function.
$label is the full name of the metric.
$Metric is either an internal metric or external metric function and gives a score to the given partition (and some given data if
it's an internal metric).
$maximize is a logical value indicating if the score returned is best when maximized (TRUE)
or when minimized (FALSE).
}
\description{
Load the selected built-in metric as a 'metric-list' to be used in Tuning function or directly by the user using $Metric.
Use names(GetMetric()) to have a list of all the built-in metrics.
}
