% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{homogeneity_completeness_vmeasure}
\alias{homogeneity_completeness_vmeasure}
\title{Homogeneity, completeness and V-Measure metrics for partition membership}
\usage{
homogeneity_completeness_vmeasure(pred_partition, true_partition)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{true_partition}{"true" partition, aka ground truth}
}
\value{
The homogeneity, the completeness and the v_measure
of the two partitions.
}
\description{
External metrics.
}
\details{
Like scikitlearn, we'll assume the inputs are two vectors of categoricals,
or things that can be turned into categoricals.
}
\examples{
  x <- c(0, 0, 1, 1, 2, 2, 3, 3, 4, 4)
  y <- c(1, 1, 2, 2, 2, 2, 6, 7, 9, 9)
  homogeneity_completeness_vmeasure(x, y)

}
