% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{meila_vi}
\alias{meila_vi}
\title{Meila's variation of information index}
\usage{
meila_vi(pred_partition, true_partition)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{true_partition}{"true" partition, aka ground truth}
}
\value{
the Meila's variation of information index of the partitions
}
\description{
Meila's variation of information index
}
