% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mutual_information}
\alias{mutual_information}
\title{Calculate mutual information of two vectors}
\usage{
mutual_information(pred_partition, true_partition)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{true_partition}{"true" partition, aka ground truth}
}
\value{
the mutual information of the two partitions
}
\description{
External metric.
}
\details{
consistent with sklearn
}
