% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{normalized_mutual_information}
\alias{normalized_mutual_information}
\title{Normalized mutual information}
\usage{
normalized_mutual_information(pred_partition, true_partition)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{true_partition}{"true" partition, aka ground truth}
}
\value{
The normalized mutual information of the two partitions
}
\description{
External metric.
}
