% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_affinity_matrix}
\alias{plot_affinity_matrix}
\title{Plot clustered affinity matrix}
\usage{
plot_affinity_matrix(affinity_matrix, partition)
}
\arguments{
\item{affinity_matrix}{An affinity matrix.}

\item{partition}{A partition of the samples.}
}
\description{
To be used with SNF and ANF affinity matrices for example.
}
