% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PreProcess}
\alias{PreProcess}
\title{Pre-process the features of a given feature matrix.}
\usage{
PreProcess(data, method = c("center", "scale"), ...)
}
\arguments{
\item{data}{a feature matrix: samples in rows, features in columns.}

\item{method}{a character vector specifying the type of processing.
Possible values are "BoxCox", "YeoJohnson", "expoTrans", "center", "scale", "range",
"knnImpute", "bagImpute", "medianImpute", "pca", "ica", "spatialSign", "corr", "zv", "nzv",
and "conditionalX" (see Details of \code{\link[caret]{preProcess}})}

\item{...}{additional arguments that will be passed to \code{\link[caret]{preProcess}}}
}
\value{
the preprocessed data.
}
\description{
Use the \code{\link[caret]{preProcess}} function of the package caret to preprocess the data.
Centering and scaling it by default
}
