% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_selection.R
\name{select_features}
\alias{select_features}
\title{Selection of important features}
\usage{
select_features(data, method = c("variance", "pca", "mad"), ...)
}
\arguments{
\item{data}{a feature matrix, rows are samples and columns are features}

\item{method}{"variance", "pca",}

\item{...}{args to be passed to the selected method:
\itemize{
\item "variance": \code{cutoff}: Minimum variance to be kept.
\item "pca":
\itemize{
\item \code{min_pve}: Minimum of percentage of variance explained by the
selected principal components.
\item \code{scale}: A logical value indicating wether the data is normalized
before PCA.
}
\item "mad": \code{cutoff}: Minimum MAD to be kept.
}}
}
\value{
data with only selected features or principal components.
}
\description{
The parameters for each methods:
}
\details{
\itemize{
\item "variance": \code{cutoff}: Minimum variance to be kept.
\itemize{
\item "pca": \code{min_pve}: Minimum of percentage of variance explained by the
selected principal components.
\item "mad": \code{cutoff}: Minimum MAD to be kept.
}
}
}
