% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_selection.R
\name{select_mad}
\alias{select_mad}
\title{Mean Absolute Deviation (MAD) cutoff}
\usage{
select_mad(data, cutoff)
}
\arguments{
\item{data}{A feature matrix, rows are samples and columns are features.}

\item{cutoff}{Minimum MAD to be kept.}
}
\value{
data with only selected features.
}
\description{
Mean Absolute Deviation (MAD) cutoff
}
