% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_selection.R
\name{select_pca}
\alias{select_pca}
\title{Apply principal components analysis}
\usage{
select_pca(data, min_pve = 1, scale = TRUE)
}
\arguments{
\item{data}{A feature matrix, rows are samples and columns are features.}

\item{min_pve}{Minimum of percentage of variance explained by the selected
principal components.}

\item{scale}{A logical value indicating wether the data is normalized before PCA.}
}
\value{
A features matrix with only selected principal components.
}
\description{
Apply principal components analysis
}
