% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_selection.R
\name{select_variance}
\alias{select_variance}
\title{Variance cutoff}
\usage{
select_variance(data, cutoff)
}
\arguments{
\item{data}{A feature matrix, rows are samples and columns are features.}

\item{cutoff}{Minimum variance to be kept.}
}
\value{
data with only selected features.
}
\description{
Variance cutoff
}
