% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{silhouette_distance}
\alias{silhouette_distance}
\title{Classic silhouette index for distance matrix}
\usage{
silhouette_distance(pred_partition, distance_matrix, average = FALSE)
}
\arguments{
\item{pred_partition}{partition predicted by a method, an integer vector}

\item{distance_matrix}{a distance matrix}
}
\value{
a silhouette object
}
\description{
External metric. To be used with distance matrix and not affinity matrix!
}
\seealso{
\code{\link[cluster]{silhouette}}
}
