% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype.R
\name{subtype}
\alias{subtype}
\title{Find subtypes among the patients}
\usage{
subtype(data_list, method, ...)
}
\arguments{
\item{data_list}{a list of data matrices with continuous data of format
samples x features (with the same number of samples).}

\item{method}{a string being the name of the built-in method to be used or a
'method list' with the same format as the built-in ones.}

\item{...}{parameters to be passed to the selected \code{method}}
}
\description{
subtype will apply the selected method to find subtypes among the patients.
}
