% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subtype_anf}
\alias{subtype_anf}
\title{Subtypes using Affinity Network Fusion}
\usage{
subtype_anf(data_list, minimal_return = FALSE, cluster_number, k_affi,
  alpha_affi = 1/6, beta_affi = 1/6, k_fusion = 20,
  weigth_fusion = NULL, type_fusion = c("two-step", "one-step"),
  alpha_fusion = c(1, 1, 0, 0, 0, 0, 0, 0), spectral_type = c("rw", "sym",
  "unnormalized"), verbose_fusion = FALSE)
}
\arguments{
\item{data_list}{a list of data matrices with continuous data of format
samples x features (with the same number of samples).}

\item{minimal_return}{logical, if TRUE, the result of the function will just
be what's needed to evaluate the goodness of the partition, i.e. the
partition and the element for internal metrics.}

\item{cluster_number}{The supposed or previously infered number of clusters.}

\item{k_affi}{The number of k-nearest neighbors for the affinity matrix
calculation.}

\item{alpha_affi}{Coefficient for local diameters for affinity matrix
calculation. Default value: 1/6.}

\item{beta_affi}{Coefficient for pair-wise distance for affinity matrix
calculation. Default value: 1/6.}

\item{k_fusion}{the number of k nearest neighbors for function kNN_graph}

\item{weigth_fusion}{a list of non-negative real numbers (which will be
normalized internally so that it sums to 1) that one-to-one correspond to
the affinity matrices computed from the features matrices provided in
\code{data_list}. If not set, internally uniform weights are assigned to all
the affinity matrices.}

\item{type_fusion}{choose one of the two options "one-step" random
walk, or "two-step" random walk (for the fusion process).}

\item{alpha_fusion}{a list of eight non-negative real numbers (which will be
normalized internally to make it sums to 1). Only used when "two-step"
(default value of \code{type_fusion}) random walk is used. \code{alpha_fusion} is the
weights for eight terms in the "two-step" random walk formula (check
research paper for more explanations about the terms).
Default value: (1, 1, 0, 0, 0, 0, 0, 0), i.e., only use the first two terms
(since they are most effective in practice).}

\item{spectral_type}{choose one of three versions of graph Laplacian:
"unnormalized": unnormalized graph Laplacian matrix (L = D - W\code{); "rw": normalization closely related to random walk (L = I - D^(-1)*W)};
(default choice) "sym": normalized symmetric matrix
(L = I - D^(-0.5) * W * D^(-0.5)`)
For more information:
\url{https://www.cs.cmu.edu/~aarti/Class/10701/readings/Luxburg06_TR.pdf}}

\item{verbose_fusion}{logical(1); if true, print some information concerning
the fusion step.}
}
\value{
a result list containing:
\itemize{
\item $partition: The predicted partition
\item $element_for_metric: The name of the element in the result list.
containing the data to be used with internal metrics.
\item $affinity_fused: The fused affinity matrix returned by the function
\code{\link[ANF]{ANF}}
}
}
\description{
Subtypes using Affinity Network Fusion
}
\seealso{
\code{\link[ANF]{ANF}}, \code{\link[ANF]{affinity_matrix}},
\code{\link[ANF]{spectral_clustering}}.
}
