% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subtype_consensus_clustering}
\alias{subtype_consensus_clustering}
\title{Subtypes using ConsensusClusterPlus}
\usage{
subtype_consensus_clustering(data_list, minimal_return = FALSE,
  maxK = 3, reps = 10, pItem = 0.8, pFeature = 1,
  clusterAlg = "hc", title = "untitled_consensus_cluster",
  innerLinkage = "average", finalLinkage = "average",
  distance = "pearson", ml = NULL, tmyPal = NULL, seed = NULL,
  plot = NULL, writeTable = FALSE, weightsItem = NULL,
  weightsFeature = NULL, verbose = F, corUse = "everything")
}
\arguments{
\item{corUse}{}
}
\description{
ConsensusClusterPlus function for determing cluster number and class
membership by stability evidence. calcICL function for calculating
cluster-consensus and item-consensus.
}
\seealso{
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
