% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subtype_pins}
\alias{subtype_pins}
\title{Subtype using PINSPlus package}
\usage{
subtype_pins(data_list, minimal_return = FALSE, return_stage_2 = TRUE,
  k_max = 5, agreement_cutoff = 0.5, verbose = T, ...)
}
\arguments{
\item{data_list}{a list of data matrices with continuous data of format
samples x features (with the same number of samples).}

\item{minimal_return}{logical, if TRUE, the result of the function will just
be what's needed to evaluate the goodness of the partition, i.e. the
partition and the element for internal metrics.}

\item{return_stage_2}{logical to return the partition of the stage 2 of the
PINS method's workflow.}

\item{k_max}{The maximum number of clusters tested (from 2 to k_max).
Default value is 5.}

\item{agreement_cutoff}{Agreement threshold to be considered consistent.
Default value is 0.5.}

\item{verbose}{Set it to TRUE of FALSE to get more or less details
respectively.}

\item{...}{these arguments will be passed to PerturbationClustering
algorithm. See details for more information.}
}
\value{
a result list containing:
\itemize{
\item $partition: The predicted partition
\item $element_for_metric: The name of the element in the result list.
containing the data to be used with internal metrics.
\item $dataTypeResult: A list of results for individual data type.
Each element of the list is the result of
\code{\link[PINSPlus]{PerturbationClustering}}
for the corresponding data matrix provided in dataList.
}
}
\description{
Perform subtyping using multiple types of data
}
\details{
SubtypingOmicsData implements the Subtyping multi-omic data that are based on
Perturbaion clustering algorithm of Nguyen, et al (2017). The input is a
list of data matrices where each matrix represents the molecular
measurements of a data type. The input matrices must have the same number
of rows. SubtypingOmicsData aims to find the optimum number of subtypes
and location of each sample in the clusters from integrated input data
dataList through two processing stages:
\itemize{
\item Stage I: The algorithm first partitions each data type using the function
PerturbationClustering. It then merges the connectivities across data
types into similarity matrices. Both kmeans and similarity-based
clustering algorithms - partitioning around medoids pam are used to
partition the built similarity. The algorithm returns the partitioning
that agrees the most with individual data types.
\item Stage II: The algorithm attempts to split each discovered group if there
is a strong agreement between data types, or if the subtyping in Stage I
is very unbalanced.
}
}
\seealso{
\code{\link[PINSPlus]{PerturbationClustering}},
\code{\link[PINSPlus]{SubtypingOmicsData}}.
}
