% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subtype_snf}
\alias{subtype_snf}
\title{Subtypes patients using Similarity Network Fusion}
\usage{
subtype_snf(data_list, minimal_return = FALSE, cluster_number, K = 20,
  alpha = 0.5, t = 20, spectral_clust_type = 3, just_fuse = FALSE)
}
\arguments{
\item{data_list}{a list of data matrices with continuous data of format
samples x features (with the same number of samples).}

\item{minimal_return}{logical, if TRUE, the result of the function will just
be what's needed to evaluate the goodness of the partition, i.e. the
partition and the element for internal metrics.}

\item{cluster_number}{The supposed or previously infered number of clusters.}

\item{K}{Number of neighbors in K-nearest neighbors part of the algorithm of
fusion and for the computation of the affinity matrix (same parameter for
both process).}

\item{alpha}{Variance for the local model (for the Gaussian kernel of the
affinity matrix). Recommended values are between 0.3 and 0.8.}

\item{t}{Number of iterations for the diffusion process.}

\item{spectral_clust_type}{The type of spectral clustering, see
\code{\link[SNFtool]{spectralClustering}} for more information.}

\item{just_fuse}{wetherto just integrate matrices and return the fused matrix
or not}
}
\value{
a result list containing:
\itemize{
\item $partition: The predicted partition
\item $element_for_metric: The name of the element in the result list.
containing the data to be used with internal metrics.
\item $affinity_fused: The fused affinity matrix returned by the function
\code{\link[SNFtool]{SNF}}.
}
}
\description{
Subtypes patients using Similarity Network Fusion
}
\seealso{
\code{\link[SNFtool]{SNF}},
\code{\link[SNFtool]{spectralClustering}},
\code{\link[SNFtool]{affinityMatrix}}.
}
