% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{synth_based_on_data}
\alias{synth_based_on_data}
\title{Generate synthetic data}
\usage{
synth_based_on_data(data.support, structure.type = c("basic", "moClus"),
  separation = 2)
}
\arguments{
\item{data.support}{a list of features matrix that are used to generate the data}

\item{structure.type}{the type of structure, "moClus" is inspired from
moCluster: Identifying Joint Patterns Across Multiple Omics Data Sets,
DOI: 10.1021/acs.jproteome.5b00824}

\item{separation}{adjust the separation between clusters.
It's the space between the means of the normal distributions used to generate the
structure of the synthetic data}
}
\description{
Generate synthetic data with a known structure based on a real dataset provided by the user
}
