% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_exhaustively}
\alias{tune_exhaustively}
\title{Systematically search parameter space for best fit}
\usage{
tune_exhaustively(tune_fxn, params, maximize, seed = NULL, cores = NULL)
}
\arguments{
\item{cores}{}
}
\description{
Systematically search parameter space for best fit
}
