% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_lhc}
\alias{tune_lhc}
\title{Tune parameters by Latin Hypercube sampling}
\usage{
tune_lhc(tune_fxn, params, maximize, seed = NULL, cores = NULL)
}
\arguments{
\item{cores}{}
}
\description{
NOTE: how do we divide up space?
}
