% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{Tuning}
\alias{Tuning}
\title{Tune the methods to have the best set of parameters}
\usage{
Tuning(data.list, method, grid.support, metric = "asw.affinity",
  ground.truth = NULL, parallel = TRUE, plot = FALSE, verbose = TRUE)
}
\arguments{
\item{data.list}{a list of data matrices with continuous data of format samples x features (with the same number of samples).}

\item{method}{a string being the name of the built-in method to be used or a 'method list' with the same format as the built-in ones.}

\item{grid.support}{a list with a set of value for each parameter to be tuned. The typo has to be correct.
Use formals(method$Func) to know all the available parameters.}

\item{metric}{a string being the name of the built-in metric to be used or a 'metric list'
with the same format as the built-in ones. If the chosen metric is internal, be sure to choose a metric corresponding to the kind
of data returned by the chosen method.}

\item{ground.truth}{a factor or integer vector being a classification of the patients. It will be used with external metrics.}

\item{parallel}{logical, TRUE for parallel computing (default and recommanded)}

\item{plot}{logical, set TRUE to plot the metric evaluation of each grid point to have an overview
of the impact of the variation of the parameters on the value of the considered metric.}

\item{verbose}{logical, set FALSE to avoid printing of informations}
}
\value{
the result list.
$metric.val contain the best value of the metric among all the grid

$parameters is the corresponding set of parameters

$method.used contains the name of the tuned method, to know where the result comes from.

$ method.res contains the result of the method, i.e. a partition and the data returned by the method,
to allow external and/or internal validation by other metrics.
}
\description{
Tuning considers a list of values of parameters (grid.support) to be tested (all the combinations are tested)
and find the set of parameters that have the best value for the metric selected.
}
