% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{check_distribution}
\alias{check_distribution}
\title{Plot the distribution of the features}
\usage{
check_distribution(data)
}
\arguments{
\item{data}{a feature matrix with samples in rows and features in columns}
}
\description{
The function plots the distribution of the mean, the variance and the median absolute deviation
of the features.
}
