% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{flag_data}
\alias{flag_data}
\title{Simple quality control on sample-feature matrices}
\usage{
flag_data(m, flag_fxn = is.na)
}
\arguments{
\item{m}{a sample-feature matrix}

\item{flag_fxn}{a boolean function that detects problematic data}
}
\value{
a summary of how the matrix is affected by problematic data
}
\description{
Many methods have restrictions on the data they can use (most notably
missing data). This is a simple function for assessing the quality of
input data. By default it looks for missing (NA) data although the
the trigger condition can be modified.
}
