% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_synth_data}
\alias{generate_synth_data}
\title{Generate 'multi-omic' synthetic data matrices  for validation}
\usage{
generate_synth_data(type = c("gaussian", "uniform", "structured"),
  n_samples = NULL, n_features = NULL, n_layers = NULL,
  support_data_list = NULL, separation = 2)
}
\arguments{
\item{type}{type of data, "gaussian" is a homogeneous group of patients,
made with the gaussian distribution. "structured" is made as the
simulated data is created in 10.1021/acs.jproteome.5b00824}

\item{n_samples}{the number of samples (rows)}

\item{n_features}{a list of the numbers of features for each matrix}

\item{n_layers}{integer. The number of layers (features matrix) to be
generated}

\item{support_data_list}{a list of features matrix that are used to generate
the data}

\item{separation}{double. The value of separation between clusters:
the space between the means of the normal distributions used to
generate the
structure of the synthetic data.}
}
\value{
'multi-omic' synthetic data matrices for validation
}
\description{
For validation and testing multi-omic / integrative methods, we need
}
