% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_synth_data}
\alias{generate_synth_data}
\alias{generate_synth_data}
\title{Generate synthetic data matrix for validation}
\usage{
generate_synth_data(dims)

generate_synth_data(dims)
}
\arguments{
\item{dims}{dimensions of matrix to generate, being the number of samples and features respectively}

\item{dims}{dimensions of matrix to generate, being the number of samples and features respectively}
}
\description{
Generate synthetic data matrix for validation

Generate random data matrix for validation
}
