% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_synth_data_multi}
\alias{generate_synth_data_multi}
\title{Generate 'multi-omic' synthetic data matrices  for validation}
\usage{
generate_synth_data_multi(n_samples, n_layers)
}
\arguments{
\item{n_samples}{the number of samples (rows) for each matrix}

\item{n_features}{a list of the numbers of features for each matrix}
}
\description{
For validation and testing multi-omic / integrative methods, we need
}
