% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetMethod}
\alias{GetMethod}
\title{Load built-in methods}
\usage{
GetMethod(method = NULL)
}
\arguments{
\item{method}{the name of the built-in method to be used}
}
\value{
a method list
}
\description{
Load built-in methods
}
