% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetMetric}
\alias{GetMetric}
\title{Load built-in metrics}
\usage{
GetMetric(metric = NULL)
}
\arguments{
\item{metric}{the metric to be used. use names(GetMetric()) to see all the metrics available.}
}
\value{
a metric list
}
\description{
Load built-in metrics
}
