% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{MetricValues}
\alias{MetricValues}
\title{An overview of the results with all available metrics}
\usage{
MetricValues(tuning.result, metrics.names = NULL, ground.truth = NULL,
  print = T, plot = T)
}
\arguments{
\item{tuning.result}{result from Tuning function}

\item{metrics.names}{a character vector with the name of the values to evaluate the result. If NULL, it uses all the built-in metrics}

\item{ground.truth}{a factor or integer vector being a classification of the patients. It will be used with external metrics.}

\item{print}{logical, to print the values of metrics in the console.}

\item{plot}{logical, to plot silhouette graph}
}
\value{
a data.frame with the value of metrics.
}
\description{
MetricValues use the result of the tuning and diagnoses it using a list of selected metric. It's useful to see how other metrics
evaluate the result tuned with only one given metric which is subject to biases.
}
