% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{silhouette_affinity}
\alias{silhouette_affinity}
\title{Silhouette index for affinity matrix}
\usage{
silhouette_affinity(pred_partition, affinity_matrix, average = FALSE)
}
\arguments{
\item{pred_partition}{partition predicted by a method}

\item{affinity_matrix}{an affinity matrix}
}
\value{
a silhouette object
}
\description{
External metric. To be used with affinity matrix and not distance matrix!
}
